Replaces uTP library with async-std-utp
The current libutp-rs uses a C lib and has some use of libc that are only available on Unix platforms.
This did not compile on Windows, and this PR proposes a replacement lib on top of `utp` crate, which is written in Rust.

The crate has been integrated with async-std, so there is no need to compat with tokio